Work around "Could not read from progress file." in CI tests
I think this is because the tests run make, via targets
generated by cmake. One or the other is trying to read
a progress file and the two threads are clashing.

The source of "Could not read from progress file."
appears to be
https://github.com/Kitware/CMake/blob/b3f1a292266099865e0896be9fd0d40803bda7e0/Source/cmcmd.cxx#L1509-L1511

Running with a single thread should resolve
the contention issue pending some better soloution.