[Thumb] Disable loadbinaries demo at -O3 UBSAN
My explanation in 7ddee92be0ecf5fad5bac3e88ab354a9eb871562
was incorrect. The detail is ok but what I was missing
was that this issue only happened with UBSAN enabled and
that we were running out of RAM not ROM.

We have space in ROM, how much I'm not sure but enough
for the copy of handle_format_char. What was causing the
issue was an increase in .data size when UBSAN was enabled.

$ ./bloaty ../ARMMultiTasking/print_cpp_with_ubsan -d fullsymbols --
../ARMMultiTasking/print_cpp_without_ubsan
    FILE SIZE        VM SIZE
 --------------  --------------
  <...>
  [NEW] +4.27Ki  [NEW] +4.27Ki    [section .data]
   +42% +3.94Ki  [ = ]       0    [section .rel.debug_info]
 +14e2% +3.62Ki  [ = ]       0    [section .rel.text]
  +241% +3.43Ki  +243% +3.28Ki    sprintf
  +247% +3.21Ki  +279% +3.12Ki    handle_format_char(int*, char const**,
std::__va_list, PrintOutput const&) (.constprop.0)
  <...>
  +108%    +176  +163%    +176    printf
  +119%    +176  +191%    +176    vprintf
  <...>

(generated with https://github.com/google/bloaty)

The baseline is -O3 without UBSAN, so any + is what UBSAN adds.
Various functions increase in size, which is expected.

What I didn't expect was 4.27k of .data. I don't know the
implementation details of UBSAN so my guess is it's tracking
values over time, or using the space to allocate the objects
the handlers get when there's an error.
(I wasn't able to find any specific symbols or markers within
this new 4k of data)

Anyway, the "fix" is to not run loadbinaries with UBSAN at O3.
We get most of the value from the other settings and other
platforms will still run the demo so nothing much is lost.

(We could just use more than 64k of RAM, QEMU allows it.
However there's a good chance that would come back to bite
us with future QEMU releases)

Note: this demo is already disabled for LTO