Reduce size of heap tags
Poking around with bloaty I saw that block_tags
was 1024 bytes on AArch64. Since size_t is 64 and
the struct was padded to be a multiple of 8, 16*64
blocks = 1024.

To reduce this pack two 16 bit fields into the tag,
one for thread ID and one for the tag. They should
have enough range for what we're doing.
(also this makes the size of block_tags the same
across platforms)