Don't set thread state to init on stack check failure
I always intended to do some special case where
if the yield asm saw init it wouldn't save the current
thread.

It turns out that I did that accidentally when we switched
to saving the whole context on kernel entry.

This means that when you call load_next_thread() you're
actually only loading next. Not saving current first
like it was before.