extern variables in kernel/thread.h
This prevents them being duplicated across
objects. This isn't usually a problem since
gcc defaults to -fcommon which allows this.
(they're merged later on)

https://gcc.gnu.org/onlinedocs/gcc/Code-Gen-Options.html

However trying out clang-11 I found that that
has changed the default to -fno-common.

https://releases.llvm.org/11.0.0/tools/clang/docs/ReleaseNotes.html

Since we don't intend for them to be duplicated,
add -fno-common to compile options so we are
required to use extern.