Add document describing how to build with clang
It's nothing too complicated but I know last time
I tried with clang-9 it was a pain. This is much
easier.

I'm not going to try supporting it properly at this
time. Just adding this document for reference, in
case we want to try some fun clang only option.

Adding "proper" (as proper as our weird cmake setup
can be) support is very possible.
(it was taken out along with the Linux build)

The one new thing is you'd need to symlink the
include path in CI. (though you could do that all
the time, and not harm the gcc builds)

However I don't see a need to complicate CI at
this point.