Add TODOs for get/set_thread_property using E_INVALID_ARGS errno
Until I figure out a demo to actually check the behaviour in, stick with
the assert.

In future some kind of "strict errno" or "logging errno" mode
might be useful to help debugging. We would check errno before
returning from each syscall and log/assert on it.
(I guess we'd have to clear it on entry to each syscall too)