Add kernel option to log failed syscalls that set errno
This should help debugging demos that fail "silently".
Especially as we move away from asserts to using errno more.

Since getting the address of errno is itself a syscall
(get_thread_property) we have to be careful of how
we handle errno during the syscall.

* Save the old errno, set errno to 0.
* Do the syscall.
* If errno is now non zero, log it and
  do *not* restore errno.
* If it is still zero, restore errno to
  the old value.

This handles a situation like:
assert(!<syscall>(...));
assert(errno == E_<whatever>);

The first syscall sets errno to non zero.
Then in reading errno, we do another syscall.
So if we didn't restore the old errno after the
second syscall the assertion would fail.