Simplify UserPointer usage
Describe properties as classes so you only need one
argument to the template functions.

Define a set_value for when T != the expected type
and =delete it. This means that if you try to set_value
with something that implicitly converts to the correct
type, it will still be an error.

(if we allow it, it's not the end of the world since
it casts to the right size, but I'd rather require
you to choose)

Along the way change userspace's set_thread_name
so we give the kernel the const char* directly
instead of a const char**.

This was potentially bypassing an E_INVALID_ARGS
check and also being a pain for the templates.