Add user space thread info to avoid common syscalls
This is inspired by the VDSO on Linux[0] but a lot
simpler.

Instead of doing syscalls to:
* get the kernel config
* get the current thread id
* get/set errno
* get the current thread name

We just have a structure user_thread_info that can
be read from userspace. This structure has the
thread id and name, the kernel config (all read only)
and errno (read/write).
(a macro is added so kernel code sees all fields
as read/write)

It is updated whenever syscalls might invalidate
it and swapped when we switch to a new thread.

For errno the kernel will set errno directly in the user
thread info. When the thread is swapped errno will be copied
back to the kernel's thread struct.

All this means that log_event no longer does any syscalls
and any syscall that might set errno only does that
one syscall. (not two to get errno afterwards)

Since we can do that, the following have been removed:
* syscall get_kernel_config
* TPROP_ID
* TPROP_ERRNO_PTR
(TPROP_NAME remains so you can read *other* thread's
names)

Since we don't use the MMU, we could have given users
direct access to the thread structs. However I'd like
to at least pretend there is some memory separation.
(also it will help potential future memory protection)

On this theme, errno ptr was a "security issue" in the
sense that it's a pointer directly into kernel data.
So this also "solves" that issue.

[0] https://man7.org/linux/man-pages/man7/vdso.7.html