Reverse use of underflow/overflow terms in stack checking
My use was inconsistent and made the stackcheck
demo rather confusing.

Stack pointer starts at a high address and
the stack grows down. Therefore:

Overflow - when the stack pointer is less
then the low address bound of the stack
(the classic recursion issue)

Underflow - when the stack pointer is greater
than the high address bound of the stack
(returning from the first and only function,
or writing off of the end of a buffer within
the first stack frame)

(there are some places where overflow is used
for both, but we have two canaries so we need
to differentiate)