Simplify check_stack logic
The old process was:
* call check_stack
* if current stack is invalid jump into load_next_thread
* then we're in assembly which checks next_thread
  * if it's NULL then call the scheduler
  * otherwise load next_thread

This is essentially a goto and as the TODO said,
was leaving behind a bit of a mess.

It's also a (potential) bug. If next_thread is not NULL
then we'd skip the housekeeping we do when loading
a new thread. (page swaps etc.)

I don't think that was happening because the yield
syscall doesn't set next directly, it just has a
parameter for the next thread.

If we expanded the stack checking in future it was
going to bite us.

Anyway, the new behaviour:

* check_stack
  * either it invalidates current
  * or it doesn't
* return to k_do_yield
* this sets next = to, carrying on as normal
* do_scheduler, which does the housekeeping

So A: we don't do this weird long jump into assembly,
and B: the normal flow of the syscall is the same.
Only difference being that the yielding thread
will never be run again.