Apply check_stack to both yield types
Previously it only applied to YIELD_ANY.

So if you have a bad stack and yield to
yourself (which I was surprised is possible)
if would return false since your target thread
(yourself) isn't valid.
But it returns you back into the bad stack!

To fix this put check_stack at the start of
k_do_yield. Then if we're doing a YIELD_TO
and the current thread had a stack issue,
we will not return to it if we can't yield
to its target.

Instead, convert the call into a YIELD_ANY
and continue. The scheduler won't pick
the current thread as its invalid.

The only thing that changes is that the
yield_to in the current thread never returns.
Which is fine, since we're trashing it anyway.