Apply stack checks to all syscalls
If we're ok just completely disregarding
any data from a thread with an invalid stack,
it's fine just to convert their syscalls to
YIELD_ANYs.

Changed one of the threads in the demo to
syscall for set_thread_name to demonstrate
that the check doesn't just apply to yields
anymore.

(there is still direct yields and timer switches
but that's for another time)