Split the implementation of putchar from printf and add serial driver
In this commit, we add the following changes to seperate the hardware
implementation of putchar from SerialPrintOutput class.

* Added a new interface `SerialPort`
* Added SERIAL driver setting to choose serial port driver
* Added driver for the QEMU virt serial port
* Moved UART_BASE into that driver out of CMake config

If one wants to replace the implementation from virt to another
one, the things to do are to create an init function and a putchar
function, and to init a SerialPort object in the driver.

Then set SERIAL_DRIVER to the path of the driver file.