Build all ports with our own minimal libc
This takes Oscar's lib functions for the raspi4 port
and extends them to be used with all ports.

So now BK is actually freestanding. Before we compiled
with freestanding but then linked the libc anyway.

Since strict-align is an AArch64 only option I've
left it just on the raspi4 port. Virt AArch64 doesn't
need to care.