Only yield to parents/children that can b scheduled
Messing with the shell demo I got this:
$ ps
| shell (0)
   State | suspended (2)
  Child  | ps (2)
| run (1)
   State | suspended (2)
  Parent | shell (0)
  Child  | ps (2)
| ps (2)
   State | running (1)
  Parent | shell (0)
| echo (3)
   State | finished (4)
  Parent | ps (2)

After running "run run echo foo". What should have happened is
echo, run and run should have all finished but the first one
was left suspended.

Turns out, once echo finished it was returning to run in thread 2,
which then happens to first check if it can yield to its child.
(which I'm not sure is a great idea but anyway)
It didn't check that the child was schedulable so it tried to use
it and ended up walking the threads until it found 0 which is the shell.

That left thread 1 as the initial run suspended forever since
the shell would never yield to it again.

To fix this check that a thread has a valid ID for parent/child
and that that thread ID can be scheduled, before picking that index.

It is true that we do this schedulable check while walking the threads
as well, but by that point we're out of the parent-child structure.

(and again, yielding to the child when the parent finishes seems
like a very suspect way to go)