Enable loading elf files when LTO is enabled
This is done by adding a new macro BK_EXPORT.

When elf loading is enabled BK_EXPORT is used
to mark symbols we want to retain despite
them not being referenced in the kernel itself.
(or being referenced in some way that can be optimised
out leaving no symbol to link against)

This macro expands to attribute used when elf
loading is enabled and to nothing if it ins't.

This means that for demos that don't set CODE_PAGE_SIZE,
they still get all the unused functions removed by LTO.

This change also enables the PIE elf demo for LTO,
though that uses a hardcoded symbol lookup table to
link at load time. This is in need of an update too
but not changed here.