Change add_thread flags to a pointer to flags struct
Doing some other things I found it very confusing
how sometimes we have uint16_t for the flags, then we
have to shift it around to use it in some places.

I like the idea of only being able to remove, but let's
remove - the need to shift it about. By adding a new
struct ThreadFlags and passing a pointer to one of those
to add_thread.

(pointer to because I'm not sure how all the platforms
will decide to register allocate it)

This struct means it's always a uint16_t and we can use
it directly and remove the _NO_ macros.